taskmanager/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
bootstrap4==0.1.0
distlib==0.3.2
Django==3.2.4
django-bootstrap-datepicker==1.4.0
django-bootstrap4==3.0.1
django-crispy-forms==1.12.0
django-friendship==1.9.1
filelock==3.0.12
install==1.3.4
Pillow==8.2.0
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
virtualenv==20.4.7
